#1514 Moving to PurchasePlugin class
